+++ src/sak/resource_pool.hpp
@@ -115,7 +115,7 @@ namespace sak
 
             uint32_t free() const
                 {
-                    return m_free_list.size();
+                    return (uint32_t)m_free_list.size();
                 }
 
             void recycle(const value_ptr &resource)
